+++ stable_baselines3/common/logger.py
@@ -399,167 +399,19 @@ def make_output_format(_format: str, log_dir: str, log_suffix: str = "") -> KVWr
 
 
 # ================================================================
-# API
-# ================================================================
-
-
-def record(key: str, value: Any, exclude: Optional[Union[str, Tuple[str, ...]]] = None) -> None:
-    """
-    Log a value of some diagnostic
-    Call this once for each diagnostic quantity, each iteration
-    If called many times, last value will be used.
-
-    :param key: save to log this key
-    :param value: save to log this value
-    :param exclude: outputs to be excluded
-    """
-    Logger.CURRENT.record(key, value, exclude)
-
-
-def record_mean(key: str, value: Union[int, float], exclude: Optional[Union[str, Tuple[str, ...]]] = None) -> None:
-    """
-    The same as record(), but if called many times, values averaged.
-
-    :param key: save to log this key
-    :param value: save to log this value
-    :param exclude: outputs to be excluded
-    """
-    Logger.CURRENT.record_mean(key, value, exclude)
-
-
-def record_dict(key_values: Dict[str, Any]) -> None:
-    """
-    Log a dictionary of key-value pairs.
-
-    :param key_values: the list of keys and values to save to log
-    """
-    for key, value in key_values.items():
-        record(key, value)
-
-
-def dump(step: int = 0) -> None:
-    """
-    Write all of the diagnostics from the current iteration
-    """
-    Logger.CURRENT.dump(step)
-
-
-def get_log_dict() -> Dict:
-    """
-    get the key values logs
-
-    :return: the logged values
-    """
-    return Logger.CURRENT.name_to_value
-
-
-def log(*args, level: int = INFO) -> None:
-    """
-    Write the sequence of args, with no separators,
-    to the console and output files (if you've configured an output file).
-
-    level: int. (see logger.py docs) If the global logger level is higher than
-                the level argument here, don't print to stdout.
-
-    :param args: log the arguments
-    :param level: the logging level (can be DEBUG=10, INFO=20, WARN=30, ERROR=40, DISABLED=50)
-    """
-    Logger.CURRENT.log(*args, level=level)
-
-
-def debug(*args) -> None:
-    """
-    Write the sequence of args, with no separators,
-    to the console and output files (if you've configured an output file).
-    Using the DEBUG level.
-
-    :param args: log the arguments
-    """
-    log(*args, level=DEBUG)
-
-
-def info(*args) -> None:
-    """
-    Write the sequence of args, with no separators,
-    to the console and output files (if you've configured an output file).
-    Using the INFO level.
-
-    :param args: log the arguments
-    """
-    log(*args, level=INFO)
-
-
-def warn(*args) -> None:
-    """
-    Write the sequence of args, with no separators,
-    to the console and output files (if you've configured an output file).
-    Using the WARN level.
-
-    :param args: log the arguments
-    """
-    log(*args, level=WARN)
-
-
-def error(*args) -> None:
-    """
-    Write the sequence of args, with no separators,
-    to the console and output files (if you've configured an output file).
-    Using the ERROR level.
-
-    :param args: log the arguments
-    """
-    log(*args, level=ERROR)
-
-
-def set_level(level: int) -> None:
-    """
-    Set logging threshold on current logger.
-
-    :param level: the logging level (can be DEBUG=10, INFO=20, WARN=30, ERROR=40, DISABLED=50)
-    """
-    Logger.CURRENT.set_level(level)
-
-
-def get_level() -> int:
-    """
-    Get logging threshold on current logger.
-    :return: the logging level (can be DEBUG=10, INFO=20, WARN=30, ERROR=40, DISABLED=50)
-    """
-    return Logger.CURRENT.level
-
-
-def get_dir() -> str:
-    """
-    Get directory that log files are being written to.
-    will be None if there is no output directory (i.e., if you didn't call start)
-
-    :return: the logging directory
-    """
-    return Logger.CURRENT.get_dir()
-
-
-record_tabular = record
-dump_tabular = dump
-
-
-# ================================================================
 # Backend
 # ================================================================
 
 
 class Logger(object):
-    # A logger with no output files. (See right below class definition)
-    #  So that you can still log to the terminal without setting up any output files
-    DEFAULT = None
-    CURRENT = None  # Current logger being used by the free functions above
+    """
+    The logger class.
+
+    :param folder: the logging location
+    :param output_formats: the list of output formats
+    """
 
     def __init__(self, folder: Optional[str], output_formats: List[KVWriter]):
-        """
-        the logger class
-
-        :param folder: the logging location
-        :param output_formats: the list of output format
-        """
         self.name_to_value = defaultdict(float)  # values this iteration
         self.name_to_count = defaultdict(int)
         self.name_to_excluded = defaultdict(str)
@@ -567,8 +419,6 @@ class Logger(object):
         self.dir = folder
         self.output_formats = output_formats
 
-    # Logging API, forwarded
-    # ----------------------------------------
     def record(self, key: str, value: Any, exclude: Optional[Union[str, Tuple[str, ...]]] = None) -> None:
         """
         Log a value of some diagnostic
@@ -626,6 +476,46 @@ class Logger(object):
         if self.level <= level:
             self._do_log(args)
 
+    def debug(self, *args) -> None:
+        """
+        Write the sequence of args, with no separators,
+        to the console and output files (if you've configured an output file).
+        Using the DEBUG level.
+
+        :param args: log the arguments
+        """
+        self.log(*args, level=DEBUG)
+
+    def info(self, *args) -> None:
+        """
+        Write the sequence of args, with no separators,
+        to the console and output files (if you've configured an output file).
+        Using the INFO level.
+
+        :param args: log the arguments
+        """
+        self.log(*args, level=INFO)
+
+    def warn(self, *args) -> None:
+        """
+        Write the sequence of args, with no separators,
+        to the console and output files (if you've configured an output file).
+        Using the WARN level.
+
+        :param args: log the arguments
+        """
+        self.log(*args, level=WARN)
+
+    def error(self, *args) -> None:
+        """
+        Write the sequence of args, with no separators,
+        to the console and output files (if you've configured an output file).
+        Using the ERROR level.
+
+        :param args: log the arguments
+        """
+        self.log(*args, level=ERROR)
+
     # Configuration
     # ----------------------------------------
     def set_level(self, level: int) -> None:
@@ -665,18 +555,15 @@ class Logger(object):
                 _format.write_sequence(map(str, args))
 
 
-# Initialize logger
-Logger.DEFAULT = Logger.CURRENT = Logger(folder=None, output_formats=[HumanOutputFormat(sys.stdout)])
-
-
-def configure(folder: Optional[str] = None, format_strings: Optional[List[str]] = None) -> None:
+def configure(folder: Optional[str] = None, format_strings: Optional[List[str]] = None) -> Logger:
     """
-    configure the current logger
+    Configure the current logger.
 
     :param folder: the save location
-        (if None, $SB3_LOGDIR, if still None, tempdir/baselines-[date & time])
+        (if None, $SB3_LOGDIR, if still None, tempdir/SB3-[date & time])
     :param format_strings: the output logging format
         (if None, $SB3_LOG_FORMAT, if still None, ['stdout', 'log', 'csv'])
+    :return: The logger object.
     """
     if folder is None:
         folder = os.getenv("SB3_LOGDIR")
@@ -689,46 +576,14 @@ def configure(folder: Optional[str] = None, format_strings: Optional[List[str]] 
     if format_strings is None:
         format_strings = os.getenv("SB3_LOG_FORMAT", "stdout,log,csv").split(",")
 
-    format_strings = filter(None, format_strings)
+    format_strings = list(filter(None, format_strings))
     output_formats = [make_output_format(f, folder, log_suffix) for f in format_strings]
 
-    Logger.CURRENT = Logger(folder=folder, output_formats=output_formats)
-    log(f"Logging to {folder}")
-
-
-def reset() -> None:
-    """
-    reset the current logger
-    """
-    if Logger.CURRENT is not Logger.DEFAULT:
-        Logger.CURRENT.close()
-        Logger.CURRENT = Logger.DEFAULT
-        log("Reset logger")
-
-
-class ScopedConfigure(object):
-    def __init__(self, folder: Optional[str] = None, format_strings: Optional[List[str]] = None):
-        """
-        Class for using context manager while logging
-
-        usage:
-        with ScopedConfigure(folder=None, format_strings=None):
-            {code}
-
-        :param folder: the logging folder
-        :param format_strings: the list of output logging format
-        """
-        self.dir = folder
-        self.format_strings = format_strings
-        self.prev_logger = None
-
-    def __enter__(self) -> None:
-        self.prev_logger = Logger.CURRENT
-        configure(folder=self.dir, format_strings=self.format_strings)
-
-    def __exit__(self, *args) -> None:
-        Logger.CURRENT.close()
-        Logger.CURRENT = self.prev_logger
+    logger = Logger(folder=folder, output_formats=output_formats)
+    # Only print when some files will be saved
+    if len(format_strings) > 0 and format_strings != ["stdout"]:
+        logger.log(f"Logging to {folder}")
+    return logger
 
 
 # ================================================================